+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - fix: missing request json
 - fix: override testsuite/testcase config verify
+- fix: only strip whitespaces and tabs, \n\r are left because they maybe used in changeset
 
 **Changed**